backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "appr-backend"
version = "1.0.0"
description = "AppR Enterprise Application Inventory System - Backend API"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Proprietary" }

dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.32.1",
    "sqlalchemy[asyncio]==2.0.36",
    "asyncpg==0.30.0",
    "alembic==1.14.0",
    "pydantic[email]==2.10.3",
    "pydantic-settings==2.7.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "authlib==1.3.2",
    "python3-saml==1.16.0",
    "celery[redis]==5.4.0",
    "redis[hiredis]==5.2.1",
    "structlog==24.4.0",
    "opentelemetry-api==1.29.0",
    "opentelemetry-sdk==1.29.0",
    "opentelemetry-instrumentation-fastapi==0.50b0",
    "opentelemetry-instrumentation-sqlalchemy==0.50b0",
    "opentelemetry-instrumentation-redis==0.50b0",
    "opentelemetry-instrumentation-celery==0.50b0",
    "opentelemetry-exporter-otlp-proto-grpc==1.29.0",
    "httpx==0.28.1",
    "python-multipart==0.0.20",
    "anyio==4.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "pytest-cov==6.0.0",
    "coverage==7.6.9",
    "ruff==0.8.4",
    "mypy==1.13.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-report=xml"

[tool.coverage.run]
source = ["app"]
omit = ["tests/*", "migrations/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[tool.ruff]
target-version = "py312"
line-length = 100
select = ["E", "W", "F", "I", "N", "UP", "ANN", "S", "B", "A", "C4", "RET", "SIM"]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]
